+++ src/types/onBoarding.type.ts
@@ -1,0 +1,7 @@
+export interface OnBoardingInput {
+  nickName: string;
+  playStyle: string[];
+  position: string[];
+  communication: string[];
+  voiceChannel: string[];
+}
